requirements.txt
discord.py==1.7.3
python-decouple==3.6
validators==0.18.2
discord.py[voice] ~= 1.7.3
yt_dlp 